+++ dev/examples/reduceMotion.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { motion, useAnimation, ReducedMotion } from "@framer"
+import { motion, useAnimation, useReducedMotion } from "@framer"
 
 const style = {
     width: 100,
@@ -9,21 +9,21 @@ const style = {
 }
 
 export const App = () => {
-    const controls = useAnimation()
+    const [isVisible, setIsVisible] = React.useState(false)
+    const isReducedMotion = useReducedMotion()
+    const transition = isReducedMotion ? { type: false } : { duration: 1 }
     const variants = {
-        visible: { opacity: 1 },
-        hidden: { opacity: 0 },
+        visible: { opacity: 1, transition },
+        hidden: { opacity: 0, transition },
     }
 
     React.useEffect(() => {
-        controls.start("visible")
+        setTimeout(() => setIsVisible(!isVisible), 1500)
     })
 
     return (
-        <ReducedMotion force>
-            <motion.div animate={controls} initial="hidden" drag>
-                <motion.div variants={variants} style={style} />
-            </motion.div>
-        </ReducedMotion>
+        <motion.div animate={isVisible ? "visible" : "hidden"} initial={false}>
+            <motion.div variants={variants} style={style} />
+        </motion.div>
     )
 }